refactoring size of main app.ts
placing responsibility into service handlers
