+++ src/services/handlers.ts
@@ -1,0 +1,7 @@
+import deleteKey from "./handlers/deleteKey.js";
+import getValue  from "./handlers/getValue.js";
+import setValue  from "./handlers/setValue.js";
+
+const handlers = {deleteKey, getValue, setValue};
+
+export default handlers;
